third_party/blink/web_tests/fast/css/invalidation/targeted-class-id-expected.txt

Check that targeted class invalidation works for id selectors.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(b1, null).backgroundColor is transparent
PASS getComputedStyle(b2, null).backgroundColor is transparent
PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS getComputedStyle(b1, null).backgroundColor is green
PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS getComputedStyle(b2, null).backgroundColor is green
PASS successfullyParsed is true

TEST COMPLETE
